Publiserer alle layers + fikser constants import
