+++ .tests-e2e/xp-test-data-app/.swcrc
@@ -21,7 +21,11 @@
     "loose": false,
     "externalHelpers": false,
     // Requires v1.2.50 or upper and requires target to be es2016 or upper.
-    "keepClassNames": false
+    "keepClassNames": false,
+    "baseUrl": "./src/main/resources",
+    "paths": {
+        "@constants": ["constants.js"],
+    }
   },
   "module": {
     "type": "commonjs"
