+++ .tests-e2e/xp-test-data-app/package.json
@@ -2,7 +2,7 @@
     "name": "navno-enonicxp-test-data",
     "version": "1.0.0",
     "scripts": {
-        "build": "npm-run-all swc-constants swc",
+        "build": "npm-run-all swc swc-constants",
         "lint": "eslint ./src/main/resources --ext=ts --quiet && tsc -p tsconfig.json --noemit",
         "swc": "cd src/main && swc resources --out-dir ../../build/resources/main --strip-leading-paths --config-file ../../.swcrc",
         "swc-constants": "swc ../../src/main/resources/lib/constants.ts --out-file ./build/resources/main/constants.js --config-file .swcrc"
