+++ .tests-e2e/xp-test-data-app/tsconfig.json
@@ -3,7 +3,7 @@
     "compilerOptions": {
         "baseUrl": "./src/main/resources",
         "paths": {
-            "constants": ["../../../../../src/main/resources/lib/constants.ts"],
+            "@constants": ["../../../../../src/main/resources/lib/constants.ts"],
             "@xp-types/*": ["../../../../../.xp-codegen/*"],
             "/lib/xp/content": ["../../../../../src/main/resources/types/xp-libs/lib-content"],
             "/lib/xp/context": ["../../../../../node_modules/@enonic-types/lib-context"],
